Download IGN 1:25000 scale layers

This function allows downloading layers at a 1:25000 scale from the Instituto Geográfico Nacional, Costa Rica, through Sistema Nacional de Información Territorial de Costa Rica SNIT

Usage

get_IGN_25k(layer_name, output_format, output_file)

Arguments

layer_name

The specific layer you want to download (e.g., IGN_5:delimitacion2017_5k).

output_format

The format in which to download the data (default is GeoJSON, application/json).

output_file

The path where the downloaded spatial data will be saved (e.g., .gpkg, .shp).

Value

a file

Examples

# Select a layer from the list (replace with your chosen layer)

# Replace with the actual layer name from the list

layer_name <- "IGN_25:contorno_25k"

# Specify your desired output file format and name

output_file <- "IGN_contorno_25k.gpkg"

# Download the selected layer

get_IGN_25k(layer_name, output_file = output_file)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
#> Reading layer `file1a8d566bed16' from data source 
#>   `/tmp/RtmpcMhv39/file1a8d566bed16.geojson' using driver `GeoJSON'
#> Simple feature collection with 384 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 156152.1 ymin: 608833.8 xmax: 658879.4 ymax: 1241118
#> Projected CRS: CR05 / CRTM05
#> Deleting source `IGN_contorno_25k.gpkg' failed
#> Writing layer `IGN_contorno_25k' to data source 
#>   `IGN_contorno_25k.gpkg' using driver `GPKG'
#> Writing 384 features with 4 fields and geometry type Polygon.
#> [1] "IGN_contorno_25k.gpkg"